fix: 'testes' removed and separated functions created for ButtonMenu.jsx file
