+++ src/components/ButtonMenu.jsx
@@ -2,6 +2,7 @@ import { Menu, MenuItem } from "@mui/material"
 import StyledButton from "../styles/StyledButton"
 import { useState } from "react"
 import { makeStyles } from "@mui/styles"
+import { useNavigate } from "react-router-dom"
 
 const useStyles = makeStyles({
   popOverRoot: {
@@ -9,9 +10,10 @@ const useStyles = makeStyles({
   },
 })
 
-function ButtonMenu() {
+function ButtonMenu({ topic }) {
   let currentlyHovering = false
   const styles = useStyles()
+  const navigate = useNavigate()
 
   const [anchorEl, setAnchorEl] = useState(null)
 
@@ -38,42 +40,92 @@ function ButtonMenu() {
     }, 50)
   }
 
-  return (
-    <div>
+  function handlePopOver(topic) {
+    const menuItems = topic.popOver.map((item, index) => {
+      return (
+        <MenuItem
+          key={index}
+          sx={{ color: "secondary.main" }}
+          onClick={() => {
+            handleClose, navigate(`/${item.href}`)
+          }}
+        >
+          {item.name}
+        </MenuItem>
+      )
+    })
+
+    return (
+      <div>
+        <StyledButton
+          sx={{
+            ":hover": { bgcolor: "primary.main", color: "secondary.main" },
+          }}
+          aria-owns={anchorEl ? "simple-menu" : undefined}
+          aria-haspopup='true'
+          onClick={handleClick}
+          onMouseOver={handleClick}
+          onMouseLeave={handleCloseHover}
+        >
+          {topic.name}
+        </StyledButton>
+        <Menu
+          disableAutoFocusItem
+          PaperProps={{
+            elevation: 0,
+            style: {
+              backgroundColor: "#7D8DBE",
+              borderRadius: 12,
+              marginTop: 3,
+            },
+          }}
+          transformOrigin={{
+            vertical: "top",
+            horizontal: "center",
+          }}
+          id='simple-menu'
+          anchorEl={anchorEl}
+          open={Boolean(anchorEl)}
+          onClose={handleClose}
+          MenuListProps={{
+            onMouseEnter: handleHover,
+            onMouseLeave: handleCloseHover,
+            style: { pointerEvents: "auto" },
+          }}
+          anchorOrigin={{ horizontal: "center", vertical: "bottom" }}
+          PopoverClasses={{
+            root: styles.popOverRoot,
+          }}
+          sx={{ color: "primary.main", border: 0, boxShadow: 0 }}
+        >
+          {menuItems}
+        </Menu>
+      </div>
+    )
+  }
+
+  function handleHref(topic) {
+    return (
       <StyledButton
         sx={{ ":hover": { bgcolor: "primary.main", color: "secondary.main" } }}
-        aria-owns={anchorEl ? "simple-menu" : undefined}
-        aria-haspopup='true'
-        onClick={handleClick}
+        onClick={() => {
+          handleClick, navigate(`/${topic.href}`)
+        }}
         onMouseOver={handleClick}
         onMouseLeave={handleCloseHover}
       >
-        Open Menu
+        {topic.name}
       </StyledButton>
-      <Menu
-        id='simple-menu'
-        anchorEl={anchorEl}
-        open={Boolean(anchorEl)}
-        onClose={handleClose}
-        MenuListProps={{
-          onMouseEnter: handleHover,
-          onMouseLeave: handleCloseHover,
-          style: { pointerEvents: "auto" },
-        }}
-        getcontentanchorel={null}
-        anchorOrigin={{ horizontal: "left", vertical: "bottom" }}
-        PopoverClasses={{
-          root: styles.popOverRoot,
-        }}
-      >
-        <MenuItem sx={{ color: "primary.main" }} onClick={handleClose}>
-          Profile
-        </MenuItem>
-        <MenuItem onClick={handleClose}>My account</MenuItem>
-        <MenuItem onClick={handleClose}>Logout</MenuItem>
-      </Menu>
-    </div>
-  )
+    )
+  }
+
+  console.log(topic)
+
+  if (topic.popOver) {
+    return handlePopOver(topic)
+  } else {
+    return handleHref(topic)
+  }
 }
 
 export default ButtonMenu
